requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cloudinary==1.29.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-writable-nested==0.7.0
future==0.18.2
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
itypes==1.2.0
jdatetime==4.1.0
Jinja2==3.1.2
MarkupSafe==2.1.1
model-bakery==1.6.0
numpy==1.23.2
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
persiantools==3.0.1
Pillow==9.2.0
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tomli==2.0.1
tornado==6.2
tzdata==2022.2
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.11
whitenoise==6.2.0
XlsxWriter==3.0.3
